Labels module to HOC func
